+++ CHANGES.md
@@ -6,6 +6,8 @@ wavesurfer.js changelog
 - Markers plugin:
   - Check for event after every add/remove (#2560)
   - Add tooltip (#2595)
+  - Cursor plugin:
+    - Fix `Crash` when `destroy triggered before ready` #2606
 
 6.3.0 (03.10.2022)
 ------------------